+++ src/bittytax/conv/parsers/__init__.py
@@ -15,6 +15,7 @@ from . import (
     blockpit,
     blockscout,
     bnktothefuture,
+    bybit,
     celsius,
     cexio,
     cgtcalculator,