+++ CHANGELOG.md
@@ -121,6 +121,7 @@
 - Conversion tool: added parser for TradeOgre exchange.
 - Conversion tool: added parser for Voyager exchange.
 - Conversion tool: added parser for LBank exchange.
+- Conversion tool: added parser for Bybit exchange.
 ### Changed
 - Conversion tool: openpyxl use read-only mode. ([#337](https://github.com/BittyTax/BittyTax/issues/337))
 - Accounting tool: openpyxl use read-only mode. ([#337](https://github.com/BittyTax/BittyTax/issues/337))